test: move out UTs to restserver package, avoid import cycle
